correct prototype for 64-bit Linux

+++ bld/plusplus/c/type.c
@@ -9378,7 +9378,7 @@ static void relocType( void *e, carve_walk_base *d )
 #undef __type_reloc_size
 }
 
-pch_status PCHRelocTypes( char *block, unsigned size )
+pch_status PCHRelocTypes( char *block, size_t size )
 {
     auto type_reloc_pch_walk type_data;
